remove duplicate k8s client code
